requirements.txt
streamlit
sentence-transformers
transformers
faiss-cpu
torch
